+++ atomic/atom/startupinit/src/startupinit.js
@@ -31,7 +31,7 @@ module.exports = async (...args) => {
       let lib = {};
 
       lib["sqlite3"] = async (...args) => {
-        const [[setting, compname], pathname, sqlite3] = args;
+        const [[setting, compname], pathname, dbengine] = args;
         let output = handler.dataformat;
         try {
           let err;
@@ -41,16 +41,59 @@ module.exports = async (...args) => {
               dbconf["engine"] = "sqlite3";
               dbconf["setting"] = setting;
               dbconf["dbgroup"] = kconconf;
-              let register = await sqlite3.register(dbconf, key, compname);
+              let register = await dbengine.register(dbconf, key, compname);
               if (register.code !== 0)
-                err += `Unable establish ${key}.db3 database connection to ${dbengine} server!`;
+                err += `Unable establish ${key}.db3 database connection to ${dbconf["engine"]} server!`;
               else {
-                let conn = await sqlite3.connector(key, compname);
+                let conn = await dbengine.connector(key, compname);
                 if (conn.code != 0) err += conn.message;
                 else {
                   connection = { ...connection, ...conn.data };
                   if (!(await conn.data[key].ischema()))
                     conn.data[key].import(join(pathname, "data", `${key}.sql`));
+                }
+              }
+            }
+          }
+          if (err)
+            throw {
+              message: "Failure to create all database log files!",
+              stack: err,
+            };
+          output.data = connection;
+        } catch (error) {
+          output = errhandler(error);
+        } finally {
+          return output;
+        }
+      };
+
+      lib["mariadb"] = async (...args) => {
+        const [[setting, compname], pathname, dbengine] = args;
+        let output = handler.dataformat;
+        try {
+          let err;
+          let connection = {};
+          for (let [kconconf, conconf] of Object.entries(setting.db)) {
+            for (let [key, dbconf] of Object.entries(conconf)) {
+              dbconf["engine"] = "mariadb";
+              dbconf["setting"] = setting;
+              dbconf["dbgroup"] = kconconf;
+              let register = await dbengine.register(dbconf, key, compname);
+              if (register.code !== 0)
+                err += `Unable establish ${key} database connection to ${dbconf["engine"]} server!`;
+              else {
+                let conn = await dbengine.connector(key, compname);
+                if (conn.code != 0) err += conn.message;
+                else {
+                  connection = { ...connection, ...conn.data };
+                  if (!(await conn.data[key].ischema(key))) {
+                    await conn.data[key].import(
+                      join(pathname, "data", `${key}.sql`)
+                    );
+                    if (!(await conn.data[key].ischema(key)))
+                      err += "Import database failure!";
+                  }
                 }
               }
             }