+++ Changelog.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Optimize sqlitesession design in utils. Completed on 2025-10-05
+- Redesign mariadb module in sqlmanager engine to integration with atom/startupinit module. Completed on 2025-10-07
 
 ### Deprecated